test: split integration suites from default package tests
