+++ packages/data-table-postgres/package.json
@@ -58,6 +58,7 @@
     "prepublishOnly": "pnpm run build",
     "test": "remix-test",
     "test:bun": "bun x --bun remix-test",
+    "test:integration": "remix-test --glob.test 'src/lib/adapter.integration.test.ts'",
     "test:coverage": "remix-test --coverage",
     "typecheck": "tsgo --noEmit"
   },
