+++ packages/data-table-sqlite/package.json
@@ -48,6 +48,7 @@
     "prepublishOnly": "pnpm run build",
     "test": "remix-test",
     "test:bun": "bun x --bun remix-test",
+    "test:integration": "remix-test --glob.test 'src/lib/adapter.integration.test.ts'",
     "test:coverage": "remix-test --coverage",
     "typecheck": "tsgo --noEmit"
   },
