+++ packages/job-data-table/package.json
@@ -51,7 +51,11 @@
     "build": "tsgo -p tsconfig.build.json",
     "clean": "git clean -fdX",
     "prepublishOnly": "pnpm run build",
-    "test": "node --disable-warning=ExperimentalWarning --test",
+    "test": "node --disable-warning=ExperimentalWarning --test ./src/lib/storage.test.ts",
+    "test:integration": "node --disable-warning=ExperimentalWarning --test ./src/lib/storage.postgres.integration.test.ts ./src/lib/storage.mysql.integration.test.ts ./src/lib/storage.sqlite.test.ts",
+    "test:integration:mysql": "node --disable-warning=ExperimentalWarning --test ./src/lib/storage.mysql.integration.test.ts",
+    "test:integration:postgres": "node --disable-warning=ExperimentalWarning --test ./src/lib/storage.postgres.integration.test.ts",
+    "test:integration:sqlite": "node --disable-warning=ExperimentalWarning --test ./src/lib/storage.sqlite.test.ts",
     "typecheck": "tsgo --noEmit"
   },
   "keywords": [
