+++ packages/job-data-table/src/lib/storage.mysql.integration.test.ts
@@ -1,5 +1,7 @@
 import { after, before, describe } from 'node:test'
 import { createDatabase } from '@remix-run/data-table'
+import { createMysqlDatabaseAdapter } from '@remix-run/data-table-mysql'
+import { createPool, type Pool } from 'mysql2/promise'
 
 import { runJobStorageContract } from '../../../job/src/lib/test/storage-contract.ts'
 
@@ -10,11 +12,9 @@ import {
   setupJobStorageSchema,
 } from './test/schema.ts'
 
-import type { Pool } from 'mysql2/promise'
-
 let integrationEnabled =
-  process.env.JOB_DATA_TABLE_INTEGRATION === '1' &&
-  typeof process.env.JOB_DATA_TABLE_MYSQL_URL === 'string'
+  process.env.DATA_TABLE_INTEGRATION === '1' &&
+  typeof process.env.DATA_TABLE_MYSQL_URL === 'string'
 
 describe('data-table job storage (mysql integration)', () => {
   let pool: Pool
@@ -25,12 +25,7 @@ describe('data-table job storage (mysql integration)', () => {
       return
     }
 
-    let [{ createMysqlDatabaseAdapter }, mysql] = await Promise.all([
-      import('@remix-run/data-table-mysql'),
-      import('mysql2/promise'),
-    ])
-
-    pool = mysql.createPool(process.env.JOB_DATA_TABLE_MYSQL_URL as string)
+    pool = createPool(process.env.DATA_TABLE_MYSQL_URL as string)
     database = createDatabase(createMysqlDatabaseAdapter(pool))
     await setupJobStorageSchema(database, DEFAULT_TEST_TABLE_PREFIX)
     await resetJobStorageSchema(database, DEFAULT_TEST_TABLE_PREFIX)
